+++ emr-serverless-spark-20230808/src/Models/ListKyuubiTokenResponseBody/data/tokens.php
@@ -9,6 +9,11 @@ use AlibabaCloud\Dara\Model;
 class tokens extends Model
 {
     /**
+     * @var string[]
+     */
+    public $accountNames;
+
+    /**
      * @var int
      */
     public $createTime;
@@ -29,6 +34,11 @@ class tokens extends Model
     public $lastUsedTime;
 
     /**
+     * @var string[]
+     */
+    public $memberArns;
+
+    /**
      * @var string
      */
     public $name;
@@ -43,10 +53,12 @@ class tokens extends Model
      */
     public $tokenId;
     protected $_name = [
+        'accountNames' => 'accountNames',
         'createTime' => 'createTime',
         'createdBy' => 'createdBy',
         'expireTime' => 'expireTime',
         'lastUsedTime' => 'lastUsedTime',
+        'memberArns' => 'memberArns',
         'name' => 'name',
         'token' => 'token',
         'tokenId' => 'tokenId',
@@ -54,12 +66,29 @@ class tokens extends Model
 
     public function validate()
     {
+        if (\is_array($this->accountNames)) {
+            Model::validateArray($this->accountNames);
+        }
+        if (\is_array($this->memberArns)) {
+            Model::validateArray($this->memberArns);
+        }
         parent::validate();
     }
 
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->accountNames) {
+            if (\is_array($this->accountNames)) {
+                $res['accountNames'] = [];
+                $n1 = 0;
+                foreach ($this->accountNames as $item1) {
+                    $res['accountNames'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (null !== $this->createTime) {
             $res['createTime'] = $this->createTime;
         }
@@ -74,6 +103,17 @@ class tokens extends Model
 
         if (null !== $this->lastUsedTime) {
             $res['lastUsedTime'] = $this->lastUsedTime;
+        }
+
+        if (null !== $this->memberArns) {
+            if (\is_array($this->memberArns)) {
+                $res['memberArns'] = [];
+                $n1 = 0;
+                foreach ($this->memberArns as $item1) {
+                    $res['memberArns'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (null !== $this->name) {
@@ -99,6 +139,17 @@ class tokens extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['accountNames'])) {
+            if (!empty($map['accountNames'])) {
+                $model->accountNames = [];
+                $n1 = 0;
+                foreach ($map['accountNames'] as $item1) {
+                    $model->accountNames[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (isset($map['createTime'])) {
             $model->createTime = $map['createTime'];
         }
@@ -113,6 +164,17 @@ class tokens extends Model
 
         if (isset($map['lastUsedTime'])) {
             $model->lastUsedTime = $map['lastUsedTime'];
+        }
+
+        if (isset($map['memberArns'])) {
+            if (!empty($map['memberArns'])) {
+                $model->memberArns = [];
+                $n1 = 0;
+                foreach ($map['memberArns'] as $item1) {
+                    $model->memberArns[$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (isset($map['name'])) {