+++ emr-serverless-spark-20230808/src/Models/ListMembersResponseBody.php
@@ -1,0 +1,122 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models;
+
+use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models\ListMembersResponseBody\members;
+
+class ListMembersResponseBody extends Model
+{
+    /**
+     * @var int
+     */
+    public $maxResults;
+
+    /**
+     * @var members[]
+     */
+    public $members;
+
+    /**
+     * @var string
+     */
+    public $nextToken;
+
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var int
+     */
+    public $totalCount;
+    protected $_name = [
+        'maxResults' => 'maxResults',
+        'members' => 'members',
+        'nextToken' => 'nextToken',
+        'requestId' => 'requestId',
+        'totalCount' => 'totalCount',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->members)) {
+            Model::validateArray($this->members);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->maxResults) {
+            $res['maxResults'] = $this->maxResults;
+        }
+
+        if (null !== $this->members) {
+            if (\is_array($this->members)) {
+                $res['members'] = [];
+                $n1 = 0;
+                foreach ($this->members as $item1) {
+                    $res['members'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->nextToken) {
+            $res['nextToken'] = $this->nextToken;
+        }
+
+        if (null !== $this->requestId) {
+            $res['requestId'] = $this->requestId;
+        }
+
+        if (null !== $this->totalCount) {
+            $res['totalCount'] = $this->totalCount;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['maxResults'])) {
+            $model->maxResults = $map['maxResults'];
+        }
+
+        if (isset($map['members'])) {
+            if (!empty($map['members'])) {
+                $model->members = [];
+                $n1 = 0;
+                foreach ($map['members'] as $item1) {
+                    $model->members[$n1] = members::fromMap($item1);
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['nextToken'])) {
+            $model->nextToken = $map['nextToken'];
+        }
+
+        if (isset($map['requestId'])) {
+            $model->requestId = $map['requestId'];
+        }
+
+        if (isset($map['totalCount'])) {
+            $model->totalCount = $map['totalCount'];
+        }
+
+        return $model;
+    }
+}