+++ emr-serverless-spark-20230808/src/Models/ListMembersResponseBody/members/roles.php
@@ -1,0 +1,122 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models\ListMembersResponseBody\members;
+
+use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models\ListMembersResponseBody\members\roles\actions;
+
+class roles extends Model
+{
+    /**
+     * @var actions[]
+     */
+    public $actions;
+
+    /**
+     * @var int
+     */
+    public $createTime;
+
+    /**
+     * @var string
+     */
+    public $description;
+
+    /**
+     * @var string
+     */
+    public $roleArn;
+
+    /**
+     * @var string
+     */
+    public $roleName;
+    protected $_name = [
+        'actions' => 'actions',
+        'createTime' => 'createTime',
+        'description' => 'description',
+        'roleArn' => 'roleArn',
+        'roleName' => 'roleName',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->actions)) {
+            Model::validateArray($this->actions);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->actions) {
+            if (\is_array($this->actions)) {
+                $res['actions'] = [];
+                $n1 = 0;
+                foreach ($this->actions as $item1) {
+                    $res['actions'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->createTime) {
+            $res['createTime'] = $this->createTime;
+        }
+
+        if (null !== $this->description) {
+            $res['description'] = $this->description;
+        }
+
+        if (null !== $this->roleArn) {
+            $res['roleArn'] = $this->roleArn;
+        }
+
+        if (null !== $this->roleName) {
+            $res['roleName'] = $this->roleName;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['actions'])) {
+            if (!empty($map['actions'])) {
+                $model->actions = [];
+                $n1 = 0;
+                foreach ($map['actions'] as $item1) {
+                    $model->actions[$n1] = actions::fromMap($item1);
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['createTime'])) {
+            $model->createTime = $map['createTime'];
+        }
+
+        if (isset($map['description'])) {
+            $model->description = $map['description'];
+        }
+
+        if (isset($map['roleArn'])) {
+            $model->roleArn = $map['roleArn'];
+        }
+
+        if (isset($map['roleName'])) {
+            $model->roleName = $map['roleName'];
+        }
+
+        return $model;
+    }
+}