+++ emr-serverless-spark-20230808/ChangeLog.md
@@ -1,3 +1,14 @@
+2025-08-20 Version: 1.15.0
+- Support API ListMembers.
+- Update API CreateSessionCluster: add request parameters body.clientToken.
+- Update API ListJobRuns: add request parameters applicationConfigs.
+- Update API ListJobRuns: add request parameters runtimeConfigs.
+- Update API ListKyuubiServices: add response parameters Body.data.kyuubiServices.$.kyuubiReleaseVersion.
+- Update API ListKyuubiSparkApplications: add response parameters Body.applications.$.exitReason.
+- Update API ListKyuubiToken: add response parameters Body.data.tokens.$.accountNames.
+- Update API ListKyuubiToken: add response parameters Body.data.tokens.$.memberArns.
+
+
 2025-08-15 Version: 1.14.0
 - Support API ListSqlStatementContents.
 - Update API ListKyuubiSparkApplications: add response parameters Body.applications.$.latestSqlStatementStatus.